renamed gem using calme case
