+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in activerecord-reset-pk-sequence.gemspec
+# Specify your gem's dependencies in activerecord_reset_pk_sequence.gemspec
 gemspec
